feat(#518): ICA: ripgrep-search — Submodule shallow fragmentation
